feat: edit user

+++ src/services/admin.ts
@@ -5,8 +5,15 @@ import { ErrorWithStatus } from '../util/errorhandler';
 import { db } from '../db/config';
 import { addAudit, getAllAudits } from '../db/audit';
 import { Logger } from '../util/logger';
-import { getAllUsers, getUserById } from '../db/users';
-import { deleteUserSchema } from '../routes/admin/schema';
+import {
+  findProfilePictureByUsername,
+  getAllUsers,
+  getUserById,
+  getUserByUsername,
+  updateProfilePicturePath,
+} from '../db/users';
+import { deleteUserSchema, editUserSchema } from '../routes/admin/schema';
+import * as fs from 'node:fs/promises';
 
 export const getUsers = async (_: Request, res: TypedResponse, next: NextFunction) => {
   try {
@@ -28,6 +35,55 @@ export const getAuditLogs = async (_: Request, res: TypedResponse, next: NextFun
   }
 };
 
+export const editUser = async (
+  req: TypedRequest<typeof editUserSchema>,
+  res: TypedResponse,
+  next: NextFunction,
+) => {
+  const adminUsername = res.locals.username;
+  const { userId } = req.params;
+  const newUser = req.body;
+  try {
+    await db.tx(async (t) => {
+      const currentUser = await getUserById(t, userId);
+      const currentPfpPath = await findProfilePictureByUsername(t, currentUser.username);
+      if (currentPfpPath && !newUser.profile_picture_path) {
+        await updateProfilePicturePath(t, { username: currentUser.username, path: undefined });
+        await fs.rm(currentPfpPath, { force: true });
+      }
+      await t.none(
+        `UPDATE users
+          SET username = $1, email = $2, email_verified_on = $3, role_id = $4, totp_enabled_on = $5
+          WHERE user_id = $6`,
+        [
+          newUser.username,
+          newUser.email || undefined,
+          newUser.email_verified_on,
+          newUser.role_id,
+          newUser.totp_enabled_on,
+          userId,
+        ],
+      );
+      await addAudit(t, {
+        username: adminUsername,
+        action: 'account_updated',
+        table_name: 'users',
+        record_id: userId,
+        old_data: { ...currentUser, password: undefined },
+        new_data: { ...currentUser, ...newUser, password: undefined },
+      });
+    });
+    res.status(200).json(success([{ name: 'user_updated', message: 'User updated successfully' }]));
+  } catch (error) {
+    if (error instanceof ErrorWithStatus) {
+      next(error);
+    } else {
+      Logger.error((error as Error).stack);
+      next(new ErrorWithStatus(500, 'admin_error', 'Unknown error while trying to edit user'));
+    }
+  }
+};
+
 export const deleteUser = async (
   req: TypedRequest<typeof deleteUserSchema>,
   res: TypedResponse,
@@ -37,7 +93,11 @@ export const deleteUser = async (
   const { userId } = req.params;
   try {
     await db.tx(async (t) => {
+      const adminUser = await getUserByUsername(t, adminUsername);
       const user = await getUserById(t, userId);
+      if (adminUser.user_id === user.user_id) {
+        throw new ErrorWithStatus(400, 'admin_error', 'You cannot delete yourself');
+      }
       await t.none('DELETE FROM users WHERE user_id = $1', user.user_id);
       await addAudit(t, {
         username: adminUsername,
@@ -49,7 +109,11 @@ export const deleteUser = async (
     });
     res.status(200).json(success([{ name: 'user_deleted', message: 'User deleted successfully' }]));
   } catch (error) {
-    Logger.error((error as Error).stack);
-    next(new ErrorWithStatus(500, 'admin_error', 'Unknown error while trying to delete user'));
+    if (error instanceof ErrorWithStatus) {
+      next(error);
+    } else {
+      Logger.error((error as Error).stack);
+      next(new ErrorWithStatus(500, 'admin_error', 'Unknown error while trying to delete user'));
+    }
   }
 };
